Remove duplicate delgators for all services
mergeDelegators' documentation says that it merges delegators avoiding
multiple same delegators for the same service. However, from
experimenting with it, duplicates were allowed, as best as I could tell.
So, I've created this small patch to ensure that duplicates are removed,
along with a covering test for it to verify the change.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>